# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Astyra Makeup Try-On API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreateSubmissionRequest {
  occasion: string;
  look: string;
  firstName?: string | null;
  email?: string | null;
  imageDataUrl?: string | null;
}

export interface Submission {
  id: number;
  occasion: string;
  look: string;
  firstName?: string | null;
  email?: string | null;
  imageDataUrl?: string | null;
  createdAt: string;
}

export interface SubmissionListResponse {
  submissions: Submission[];
  total: number;
  page: number;
  pageSize: number;
}

export interface OccasionCount {
  occasion: string;
  count: number;
}

export interface LookCount {
  look: string;
  count: number;
}

export interface AdminStats {
  totalSubmissions: number;
  submissionsToday: number;
  submissionsThisWeek: number;
  byOccasion: OccasionCount[];
  byLook: LookCount[];
  emailCaptureRate: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListSubmissionsParams = {
  page?: number;
  pageSize?: number;
};