# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Astyra Makeup Try-On API
 * OpenAPI spec version: 0.1.0
 */
import type { LookCount } from "./lookCount";
import type { OccasionCount } from "./occasionCount";

export interface AdminStats {
  totalSubmissions: number;
  submissionsToday: number;
  submissionsThisWeek: number;
  byOccasion: OccasionCount[];
  byLook: LookCount[];
  emailCaptureRate: number;
}