# lib/api-zod/src/generated/types/occasionCount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Astyra Makeup Try-On API
 * OpenAPI spec version: 0.1.0
 */

export interface OccasionCount {
  occasion: string;
  count: number;
}