# lib/api-zod/src/generated/types/submission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Astyra Makeup Try-On API
 * OpenAPI spec version: 0.1.0
 */

export interface Submission {
  id: number;
  occasion: string;
  look: string;
  firstName?: string | null;
  email?: string | null;
  imageDataUrl?: string | null;
  createdAt: Date;
}