# lib/api-zod/src/generated/types/submissionListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Astyra Makeup Try-On API
 * OpenAPI spec version: 0.1.0
 */
import type { Submission } from "./submission";

export interface SubmissionListResponse {
  submissions: Submission[];
  total: number;
  page: number;
  pageSize: number;
}